+++ pyud/__init__.py
@@ -19,8 +19,18 @@ You should have received a copy of the GNU General Public License
 along with pyud.  If not, see <https://www.gnu.org/licenses/>.
 """
 
+from collections import namedtuple
+
 from .definition import Definition
 from .client import AsyncClient, Client
 
 __author__ = "William Lee"
-__version__ = "1.0.1dev1"
+__version__ = "1.0.1rc1"
+
+VersionInfo = namedtuple(
+    'version_info', 'major minor micro releaselevel serial'
+)
+
+version_info = VersionInfo(
+    major=1, minor=0, micro=1, releaselevel='candidate', serial=1
+)